Fix critical 64-bit portability issues
Replace unsafe pointer-to-int casts with intptr_t/uintptr_t to prevent
data loss on 64-bit architectures. These casts were truncating 64-bit
pointers to 32-bit integers, causing potential crashes and data corruption.

Changes:
- src/viola/hash.c: Add stdint.h, fix printf format specifiers
- src/viola/glib_x.c: Fix Window and VObj* casts in putHashEntry_replace
- src/viola/event_x.c: Fix pointer comparisons in key event handlers
- src/viola/loader.c: Fix slotValue pointer cast
- src/libWWW/Library/Implementation/HTAnchor.c: Remove obsolete int cast

This eliminates all -Wpointer-to-int-cast warnings (203 → 0) while
maintaining compatibility with both 32-bit and 64-bit systems.

+++ src/libWWW/Library/Implementation/HTAnchor.c
@@ -104,7 +104,7 @@ PUBLIC HTChildAnchor* HTAnchor_findChild ARGS2(HTParentAnchor*, parent, CONST ch
     child = HTChildAnchor_new();
     if (TRACE)
         fprintf(stderr, "new Anchor %p named `%s' is child of %p\n", (void*)child,
-                (int)tag ? tag : (CONST char*)"", (void*)parent); /* int for apollo */
+                tag ? tag : (CONST char*)"", (void*)parent);
     HTList_addObject(parent->children, child);
     child->parent = parent;
     StrAllocCopy(child->tag, tag);
